fix: cập nhật hàm parseFirstNames (chuyển sang class random) - sửa lỗi hàm tryParseFloat. - chuyển các kiểu dữ liệu dài thành auto nếu có thể.

+++ Number.cpp
@@ -82,7 +82,7 @@ float Number::tryParseFloat(string str)
 	ReturnFlags flags;
 	float number = 0;
 
-	tie(flags, number) = parseInt(str);
+	tie(flags, number) = parseFloat(str);
 	if (!flags.successful)
 	{
 		cout << "Error code: " << flags.errorCode << endl;
